requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
yfinance==0.2.28
pymongo==4.6.0
python-multipart==0.0.6
psutil==5.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6